Upgrade client side to ampersand.js & co.
There's too much too document here as the changes were supposed to be a
POC not a rewrite, but nothing goes quite as planned so expect an update
to the README and/or wiki.

Also note that this is a work in progress and as such some tests are not
complete, some may fail, and things may just be broken in general.